blackmonk/requirements.txt
django==1.5.5
python-dateutil==1.5
geopy
boto
mailsnake
xhtml2pdf
django-mptt==0.5.4          
django-picklefield
tweepy
googlemaps
suds
gunicorn
BeautifulSoup4
gdata
httplib2
Markdown
oauth2
python-openid
stripe
pywapi
django-celery
django-celery-transactions
django-appconf
validate_email